Added v1 to routes and sending status
