+++ lib/database.js
@@ -38,5 +38,5 @@ function addQuestion (content, title, created) {
 function addAnswer (content, created, question_id)  {
   return knex('answers')
     .insert({content, created, question_id})
-    .catch((error) => console.log(error))
+.    .catch((error) => console.log(error))
 }
